# BackBoiler/src/swagger.yml
openapi: 3.0.3
info:
  title: BackBoiler Model API
  version: 1.1.0
  description: BackBoiler framework for every Django Project needs.
paths:
...
  /Connect/GetCredential/:
    post:
      operationId: Connect_GetCredential_create
      description: Get or retrun connection credential for uuid of [app, user, admin]
        typefor app get from connect; for user and admin get from user and user auth<br><br>
        <b> App Credential</b>
      summary: get connection credentials. for [app, user, admin] types.
      tags:
      - Connect
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                type:
                  type: string
                  default: app
                  enum:
                  - app
                  - admin
                  - user
                  - none
                uuid:
                  type: string
                  format: uuid
                  default: ''
                  description: if type in [app, user, admin] then need app_uuid or
                    user_uuid
              required:
              - type
              - uuid
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                return: boolean
          description: Connection processed successfully
...
  /User/GetUserbyUUID/:
    post:
      operationId: User_GetUserbyUUID_create
      description: Get user details by uuid<br><br> <b> App Credential </b>
      summary: Get user details by uuid
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  default: ''
                  format: uuid
                  description: User uuid
              required:
              - uuid
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                return: boolean
          description: ''
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid